+++ CHANGELOG.md
@@ -18,6 +18,9 @@ Contact: campanam@si.edu
 [Deprecated](#deprecated)  
 
 ## aln2baits  
+### Version 1.0.4  
+Added ability to shuffle final bait if contig end reached  
+
 ### Version 1.0  
 Updated parameters table  
 Can reverse complement baits  
@@ -77,6 +80,10 @@ Version constant added to header
 Preliminary script to generate baits from an annotation file and a reference sequence  
 
 ## baitslib  
+### Version 1.0.4  
+pyrad2baits can produce an uncollapsed reference sequence  
+Added ability to shuffle final bait if contig end reached  
+
 ### Version 1.0.3  
 Fixed bug in snp_to_baits that added 2 bp to bait lengths.  
 baitstools and baitslib version will now track package version (even if no updates to these scripts) for package clarity. Individual version numbers left only to subcommands/GUI.  
@@ -127,6 +134,10 @@ New method write_probes handles basic output
 filter_probes definition removed into separate script for access by other scripts  
 
 ## baitstools  
+### Version 1.0.4  
+pyrad2baits can produce an uncollapsed reference sequence  
+Added ability to shuffle final bait if contig end reached  
+
 ### Version 1.0.3  
 baitstools and baitslib version will now track package version (even if no updates to these scripts) for package clarity. Individual version numbers left only to subcommands/GUI.  
 
@@ -231,6 +242,10 @@ The word 'probe' changed to 'baits' in all instances for clarity
 Set default for tiling offset as 20 bp (from 60 for select_snps and 25 for tile_probes)  
 
 ## baitstoolsgui  
+### Version 1.0.4  
+pyrad2baits can produce an uncollapsed reference sequence  
+Added ability to shuffle final bait if contig end reached  
+
 ### Version 1.0  
 Linguistic complexity and maximum homopolymer length filters  
 Reverse complement option  
@@ -303,6 +318,9 @@ Versioning added to header of script
 RVM default bug fixed  
 
 ## pyrad2baits  
+### Version 1.0.4  
+pyrad2baits can produce an uncollapsed reference sequence  
+
 ### Version 0.1  
 Preliminary script to generate baits from PyRAD/ipyrad loci files  
 
@@ -327,6 +345,9 @@ Fixed bug in that reread reference sequence every time snps_to_baits run in stac
 Preliminary script to turn a Stacks summary tsv file and a reference sequence into baits  
 
 ## tilebaits  
+### Version 1.0.4  
+Added ability to shuffle final bait if contig end reached  
+
 ### Version 1.0  
 Updated parameters table  
 Can reverse complement baits